commit integration

+++ src/Controller/EquipeController.php
@@ -34,9 +34,15 @@ class EquipeController extends AbstractController
      * @route("/AfficheE", name="AfficheEquipe")
      */
     public function Affiche(EquipeRepository $repository){
-        $equipe=$repository->findAll();
-        return $this->render('equipe/Affiche.html.twig',
-            ['equipe'=>$equipe]);
+        if($this->getUser()->isAdmin())
+        {
+            $equipe=$repository->findAll();
+            return $this->render('equipe/Affiche.html.twig',
+                ['equipe'=>$equipe]);
+        }
+
+        return $this->redirectToRoute('app_login');
+
     }
     /**
      * @param EquipeRepository $repository
@@ -47,7 +53,7 @@ class EquipeController extends AbstractController
         $equipe=$paginator->paginate(
             $repository->findAll(),
             $request->query->getInt('page',1),
-            9
+            1
         );
         return $this->render('equipe/Afficheruneequipe.html.twig',
             ['equipe'=>$equipe]);
@@ -57,15 +63,20 @@ class EquipeController extends AbstractController
      * @route("/delete/{id}",name="supprimerequipe")
      */
     function Delete_joueur($id,EquipeRepository $repository){
-        $equipe=$repository->find($id);
-        $em=$this->getDoctrine()->getManager();
-        $em->remove($equipe);
-        $em->flush();
-        $this->addFlash(
-            'info',
-            'Deleted Successfully'
-        );
-        return $this->redirectToRoute('AfficheEquipe');
+        if($this->getUser()->isAdmin())
+        {
+            $equipe=$repository->find($id);
+            $em=$this->getDoctrine()->getManager();
+            $em->remove($equipe);
+            $em->flush();
+            $this->addFlash(
+                'info',
+                'Deleted Successfully'
+            );
+            return $this->redirectToRoute('AfficheEquipe');
+        }
+
+        return $this->redirectToRoute('app_login');
     }
 
     /**
@@ -74,68 +85,79 @@ class EquipeController extends AbstractController
      * @route("equipe/ajouter", name="AjouterEquipe")
      */
     function Ajouter_equipe(Request $request){
-        $equipe = new Equipe();
-        $form=$this->createForm(EquipeType::class,$equipe);
-        $form->add('ajouter',SubmitType::class);
-        $form->handleRequest($request);
-        if($form->isSubmitted() && $form->isValid()){
-            $file =$equipe->getDrapeauEquipe();
-            $uploads_directory = $this->getParameter('upload_directory');
-            $fileName = md5(uniqid()).'.'.$file->guessExtension();
-            $file->move(
-                $uploads_directory,
-                $fileName
-            );
-            $equipe->setDrapeauEquipe($fileName);
-            $em=$this->getDoctrine()->getManager();
-            $em->persist($equipe);
-            $em->flush();
-            $this->addFlash(
-                'info',
-                'Added Successfully'
-            );
-            return  $this->redirectToRoute('AfficheEquipe');
-        }
-        return $this->render('equipe/Ajouter.html.twig',[
-            'form'=>$form->createView()
-        ]);
-    }
-    /**
-     * @route ("equipe/update/{id}",name="modifierequipe")
-     */
-    function Update(EquipeRepository  $repository,$id,Request $request){
-        $equipe=$repository->find($id);
-        $form=$this->createForm(EquipeEditType::class,$equipe);
-        $form->add('Update',SubmitType::class);
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()){
-            $file = $form['drapeau_equipe']->getData();
-
-            if($file)
-            {
+        if($this->getUser()->isAdmin())
+        {
+            $equipe = new Equipe();
+            $form=$this->createForm(EquipeType::class,$equipe);
+            $form->add('ajouter',SubmitType::class);
+            $form->handleRequest($request);
+            if($form->isSubmitted() && $form->isValid()){
+                $file =$equipe->getDrapeauEquipe();
                 $uploads_directory = $this->getParameter('upload_directory');
                 $fileName = md5(uniqid()).'.'.$file->guessExtension();
                 $file->move(
                     $uploads_directory,
                     $fileName
                 );
-
                 $equipe->setDrapeauEquipe($fileName);
+                $em=$this->getDoctrine()->getManager();
+                $em->persist($equipe);
+                $em->flush();
+                $this->addFlash(
+                    'info',
+                    'Added Successfully'
+                );
+                return  $this->redirectToRoute('AfficheEquipe');
             }
-            else {
-                $equipe->setDrapeauEquipe($equipe.getDrapeauEquipe());
-            }
-            $em=$this->getDoctrine()->getManager();
-            $em->flush();
-            $this->addFlash(
-                'info',
-                'Updated Successfully'
-            );
-            return $this->redirectToRoute("AfficheEquipe");
+            return $this->render('equipe/Ajouter.html.twig',[
+                'form'=>$form->createView()
+            ]);
         }
-        return $this->render('equipe/Update.html.twig',[
-            'form'=>$form->createView()
-        ]);
+        return $this->redirectToRoute('app_login');
+
+    }
+    /**
+     * @route ("equipe/update/{id}",name="modifierequipe")
+     */
+    function Update(EquipeRepository  $repository,$id,Request $request){
+        if($this->getUser()->isAdmin())
+        {
+            $equipe=$repository->find($id);
+            $form=$this->createForm(EquipeEditType::class,$equipe);
+            $form->add('Update',SubmitType::class);
+            $form->handleRequest($request);
+            if ($form->isSubmitted() && $form->isValid()){
+                $file = $form['drapeau_equipe']->getData();
+
+                if($file)
+                {
+                    $uploads_directory = $this->getParameter('upload_directory');
+                    $fileName = md5(uniqid()).'.'.$file->guessExtension();
+                    $file->move(
+                        $uploads_directory,
+                        $fileName
+                    );
+
+                    $equipe->setDrapeauEquipe($fileName);
+                }
+                else {
+                    $equipe->setDrapeauEquipe($equipe.getDrapeauEquipe());
+                }
+                $em=$this->getDoctrine()->getManager();
+                $em->flush();
+                $this->addFlash(
+                    'info',
+                    'Updated Successfully'
+                );
+                return $this->redirectToRoute("AfficheEquipe");
+            }
+            return $this->render('equipe/Update.html.twig',[
+                'form'=>$form->createView()
+            ]);
+        }
+
+        return $this->redirectToRoute('app_login');
+
     }
     /**
      * @Route("/ListByEquipe/{id}", name="ListByEquipe", methods={"GET"})
@@ -154,31 +176,38 @@ class EquipeController extends AbstractController
      * @route("/ImprimerE", name="ImprimerEquipe", methods={"GET"})
      */
     public function Imprimer(EquipeRepository $repository): Response
-    {    // Configure Dompdf according to your needs
-        $pdfOptions = new Options();
-        $pdfOptions->set('defaultFont', 'Arial');
+    {
+        if($this->getUser()->isAdmin())
+        {
+            // Configure Dompdf according to your needs
+            $pdfOptions = new Options();
+            $pdfOptions->set('defaultFont', 'Arial');
 
-        // Instantiate Dompdf with our options
-        $dompdf = new Dompdf($pdfOptions);
-        $equipe=$repository->findAll(); //retourner toutes les objets
+            // Instantiate Dompdf with our options
+            $dompdf = new Dompdf($pdfOptions);
+            $equipe=$repository->findAll(); //retourner toutes les objets
 
 
-        // Retrieve the HTML generated in our twig file
-        $html = $this->renderView('equipe/ImprimerEquipe.html.twig',
-            ['equipe'=>$equipe]);
+            // Retrieve the HTML generated in our twig file
+            $html = $this->renderView('equipe/ImprimerEquipe.html.twig',
+                ['equipe'=>$equipe]);
 
-        // Load HTML to Dompdf
-        $dompdf->loadHtml($html);
+            // Load HTML to Dompdf
+            $dompdf->loadHtml($html);
 
-        // (Optional) Setup the paper size and orientation 'portrait' or 'portrait'
-        $dompdf->setPaper('A4', 'portrait');
+            // (Optional) Setup the paper size and orientation 'portrait' or 'portrait'
+            $dompdf->setPaper('A4', 'portrait');
 
-        // Render the HTML as PDF
-        $dompdf->render();
+            // Render the HTML as PDF
+            $dompdf->render();
 
-        // Output the generated PDF to Browser (inline view)
-        $dompdf->stream("mypdf.pdf", [
-            "Attachment" => false
-        ]);
+            // Output the generated PDF to Browser (inline view)
+            $dompdf->stream("mypdf.pdf", [
+                "Attachment" => false
+            ]);
+        }
+
+        return $this->redirectToRoute('app_login');
+
     }
 }
