+++ src/Controller/JoueurController.php
@@ -34,9 +34,15 @@ class JoueurController extends AbstractController
      * @route("/AfficheJ", name="AfficheJoueur")
      */
     public function Affiche(JoueurRepository $repository){
-        $joueur=$repository->findAll();
-        return $this->render('joueur/Affiche.html.twig',
-            ['joueur'=>$joueur]);
+        if($this->getUser()->isAdmin())
+        {
+            $joueur=$repository->findAll();
+            return $this->render('joueur/Affiche.html.twig',
+                ['joueur'=>$joueur]);
+        }
+
+        return $this->redirectToRoute('app_login');
+
     }
     /**
      * @param JoueurRepository $repository
@@ -54,15 +60,21 @@ class JoueurController extends AbstractController
      * @route("/delete/{id}",name="delete")
      */
     function Delete_joueur($id,JoueurRepository $repository){
-        $joueur=$repository->find($id);
-        $em=$this->getDoctrine()->getManager();
-        $em->remove($joueur);
-        $em->flush();
-        $this->addFlash(
-            'info',
-            'Deleted Successfully'
-        );
-        return $this->redirectToRoute('AfficheJoueur');
+        if($this->getUser()->isAdmin())
+        {
+            $joueur=$repository->find($id);
+            $em=$this->getDoctrine()->getManager();
+            $em->remove($joueur);
+            $em->flush();
+            $this->addFlash(
+                'info',
+                'Deleted Successfully'
+            );
+            return $this->redirectToRoute('AfficheJoueur');
+        }
+
+        return $this->redirectToRoute('app_login');
+
     }
 
     /**
@@ -71,68 +83,80 @@ class JoueurController extends AbstractController
      * @route("joueur/ajouter",name="AjouterJoueur")
      */
     function Ajouter_joueur(Request $request){
-        $joueur = new Joueur();
-        $form=$this->createForm(JoueurType::class,$joueur);
-        $form->add('ajouter',SubmitType::class);
-        $form->handleRequest($request);
-        if($form->isSubmitted() && $form->isValid()){
-            $file =$joueur->getPhoto();
-            $uploads_directory = $this->getParameter('upload_directory');
-            $fileName = md5(uniqid()).'.'.$file->guessExtension();
-            $file->move(
-                $uploads_directory,
-                $fileName
-            );
-            $joueur->setPhoto($fileName);
-            $em=$this->getDoctrine()->getManager();
-            $em->persist($joueur);
-            $em->flush();
-            $this->addFlash(
-                'info',
-                'Added Successfully'
-            );
-            return  $this->redirectToRoute('AfficheJoueur');
-        }
-        return $this->render('joueur/Ajouter.html.twig',[
-           'form'=>$form->createView()
-        ]);
-    }
-    /**
-     * @route ("joueur/update/{id}",name="update")
-     */
-    function Update(JoueurRepository  $repository,$id,Request $request){
-        $joueur=$repository->find($id);
-        $form=$this->createForm(JoueurEditType::class,$joueur);
-        $form->add('Update',SubmitType::class);
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()){
-            $file = $form['photo']->getData();
-
-            if($file)
-            {
+        if($this->getUser()->isAdmin())
+        {
+            $joueur = new Joueur();
+            $form=$this->createForm(JoueurType::class,$joueur);
+            $form->add('ajouter',SubmitType::class);
+            $form->handleRequest($request);
+            if($form->isSubmitted() && $form->isValid()){
+                $file =$joueur->getPhoto();
                 $uploads_directory = $this->getParameter('upload_directory');
                 $fileName = md5(uniqid()).'.'.$file->guessExtension();
                 $file->move(
                     $uploads_directory,
                     $fileName
                 );
-
                 $joueur->setPhoto($fileName);
+                $em=$this->getDoctrine()->getManager();
+                $em->persist($joueur);
+                $em->flush();
+                $this->addFlash(
+                    'info',
+                    'Added Successfully'
+                );
+                return  $this->redirectToRoute('AfficheJoueur');
             }
-            else {
-                $joueur->setPhoto($joueur.getPhoto());
-            }
-            $em=$this->getDoctrine()->getManager();
-            $em->flush();
-            $this->addFlash(
-                'info',
-                'Updated Successfully'
-            );
-            return $this->redirectToRoute("AfficheJoueur");
+            return $this->render('joueur/Ajouter.html.twig',[
+                'form'=>$form->createView()
+            ]);
         }
-        return $this->render('joueur/Update.html.twig',[
-            'form'=>$form->createView()
-        ]);
+
+        return $this->redirectToRoute('app_login');
+
+    }
+    /**
+     * @route ("joueur/update/{id}",name="update")
+     */
+    function Update(JoueurRepository  $repository,$id,Request $request){
+        if($this->getUser()->isAdmin())
+        {
+            $joueur=$repository->find($id);
+            $form=$this->createForm(JoueurEditType::class,$joueur);
+            $form->add('Update',SubmitType::class);
+            $form->handleRequest($request);
+            if ($form->isSubmitted() && $form->isValid()){
+                $file = $form['photo']->getData();
+
+                if($file)
+                {
+                    $uploads_directory = $this->getParameter('upload_directory');
+                    $fileName = md5(uniqid()).'.'.$file->guessExtension();
+                    $file->move(
+                        $uploads_directory,
+                        $fileName
+                    );
+
+                    $joueur->setPhoto($fileName);
+                }
+                else {
+                    $joueur->setPhoto($joueur.getPhoto());
+                }
+                $em=$this->getDoctrine()->getManager();
+                $em->flush();
+                $this->addFlash(
+                    'info',
+                    'Updated Successfully'
+                );
+                return $this->redirectToRoute("AfficheJoueur");
+            }
+            return $this->render('joueur/Update.html.twig',[
+                'form'=>$form->createView()
+            ]);
+        }
+
+        return $this->redirectToRoute('app_login');
+
     }
